Refactored unit tests into multiple files
