+++ tests/unit/test_secret_file_usage.py
@@ -1,0 +1,121 @@
+"""
+Tests for the secret file writing functionality in vault_anyconfig package.
+"""
+# pylint: disable=attribute-defined-outside-init
+# pylint: disable=too-few-public-methods
+# pylint: disable=no-self-use
+# pylint: disable=unused-argument
+from unittest.mock import patch, mock_open, call, Mock
+from copy import deepcopy
+from json import dumps as jdumps
+from stat import S_IRUSR, S_IWUSR
+
+import pytest
+
+from vault_anyconfig.vault_anyconfig import VaultAnyConfig
+
+from .test_config import TestConfig
+
+
+class TestWriteFile(TestConfig):
+    """
+    Tests directly accessing the save_file_from_vault method
+    """
+
+    def setup(self):
+        super().setup()
+        self.file_contents = "secret_string_to_write"
+
+        self.file_path = "/some/file/secret"
+        self.secret_path = "/secret/acme/cert"
+
+    @patch("vault_anyconfig.vault_anyconfig.chmod")
+    @patch("builtins.open", new_callable=mock_open)
+    @patch("vault_anyconfig.vault_anyconfig.Client.read")
+    def test_write_new_file(self, mock_hvac_client_read, mock_open_handle, mock_chmod):
+        """
+        Retrieves a string from a (mock) HVAC client and writes it to a (mock) file.
+        """
+        mock_hvac_client_read.return_value = {"data": {"file": self.file_contents}}
+
+        self.client.save_file_from_vault(self.file_path, self.secret_path, "file")
+
+        mock_hvac_client_read.assert_called_once_with(self.secret_path)
+
+        mock_open_handle.assert_called_once_with(self.file_path, "w")
+        mock_open_handle().write.assert_called_once_with(self.file_contents)
+
+        mock_chmod.assert_called_once_with(self.file_path, S_IRUSR)
+
+    @patch("vault_anyconfig.vault_anyconfig.chmod")
+    @patch("builtins.open", new_callable=mock_open)
+    @patch("vault_anyconfig.vault_anyconfig.isfile")
+    @patch("vault_anyconfig.vault_anyconfig.Client.read")
+    def test_write_existing_file(
+        self, mock_hvac_client_read, mock_isfile, mock_open_handle, mock_chmod
+    ):
+        """
+        Tests that chmod is called twice if the secret file already existed on disk
+        """
+        mock_isfile.return_value = True
+        mock_hvac_client_read.return_value = {"data": {"file": self.file_contents}}
+
+        self.client.save_file_from_vault(self.file_path, self.secret_path, "file")
+
+        mock_hvac_client_read.assert_called_once_with(self.secret_path)
+
+        mock_open_handle.assert_called_once_with(self.file_path, "w")
+        mock_open_handle().write.assert_called_once_with(self.file_contents)
+
+        chmod_calls = [call(self.file_path, S_IWUSR), call(self.file_path, S_IRUSR)]
+        mock_chmod.assert_has_calls(chmod_calls, any_order=False)
+
+    @patch("vault_anyconfig.vault_anyconfig.chmod")
+    @patch("builtins.open", new_callable=mock_open)
+    @patch("vault_anyconfig.vault_anyconfig.isfile")
+    @patch("vault_anyconfig.vault_anyconfig.Client.read")
+    def test_file_with_write_perm_fail(
+        self, mock_hvac_client_read, mock_isfile, mock_open_handle, mock_chmod
+    ):
+        """
+        Tests that a warning is thrown if chmod is unable to set write permission on the secret file on disk
+        """
+        mock_isfile.return_value = True
+        mock_hvac_client_read.return_value = {"data": {"file": self.file_contents}}
+
+        mock_chmod.side_effect = PermissionError(Mock(return_value="IOError 13"))
+
+        with pytest.warns(UserWarning):
+            self.client.save_file_from_vault(self.file_path, self.secret_path, "file")
+
+        mock_hvac_client_read.assert_called_once_with(self.secret_path)
+
+        mock_open_handle.assert_called_once_with(self.file_path, "w")
+        mock_open_handle().write.assert_called_once_with(self.file_contents)
+
+        chmod_calls = [call(self.file_path, S_IWUSR)]
+        mock_chmod.assert_has_calls(chmod_calls, any_order=False)
+
+    @patch("vault_anyconfig.vault_anyconfig.chmod")
+    @patch("builtins.open", new_callable=mock_open)
+    @patch("vault_anyconfig.vault_anyconfig.Client.read")
+    def test_file_with_read_perm_fail(
+        self, mock_hvac_client_read, mock_open_handle, mock_chmod
+    ):
+        """
+        Tests that a warning is thrown if chmod is unable to set read-only permissions on a secret file on disk
+        """
+        mock_hvac_client_read.return_value = {"data": {"file": self.file_contents}}
+
+        mock_chmod.side_effect = PermissionError(Mock(return_value="IOError 13"))
+
+        with pytest.warns(UserWarning):
+            self.client.save_file_from_vault(self.file_path, self.secret_path, "file")
+
+        mock_hvac_client_read.assert_called_once_with(self.secret_path)
+
+        mock_open_handle.assert_called_once_with(self.file_path, "w")
+        mock_open_handle().write.assert_called_once_with(self.file_contents)
+
+        chmod_calls = [call(self.file_path, S_IRUSR)]
+        mock_chmod.assert_has_calls(chmod_calls, any_order=False)
